Add copy to clipboard for code and pre elements
Why:

When reading documents with code snippets it's common to want to copy
the content of the code snippet into the clipboard.

This commit:

Introduces the functionality for copying any <code> or <pre> tag into
the clipboard when clicked. We update the styling to make the elements
appear clickable and add a 'copied!' flash message when something is
copied into the clipboard.