Introduce build_html script
Why:

We would like to automate the task of 'prepping' the html with the
'with-prose' class name and the tags for the css and clipboard
functionality.

This commit:

Introduces a build_html.js script which will prep all the html files in
the build directory. We include the script in our build process.